# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Kavach Parental Control API
  description: API for the Kavach parental control application
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
security:
  - bearerAuth: []
  - sessionCookie: []
...
paths:
...
  /children:
    get:
      tags: [Children]
      summary: List children
      operationId: listChildren
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Paginated list of children
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Child"
                      pagination:
                        $ref: "#/components/schemas/Pagination"
    post:
      tags: [Children]
      summary: Add a child
      operationId: createChild
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                birth_date:
                  type: string
                  format: date
      responses:
        "201":
          description: Child created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
  /children/{childId}:
    parameters:
      - $ref: "#/components/parameters/ChildId"
    get:
      tags: [Children]
      summary: Get child by ID
      operationId: getChild
      responses:
        "200":
          description: Child details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "404":
          description: Child not found
    patch:
      tags: [Children]
      summary: Update child
      operationId: updateChild
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                birth_date:
                  type: string
                  format: date
      responses:
        "200":
          description: Child updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
    delete:
      tags: [Children]
      summary: Remove child (owner only)
      operationId: deleteChild
      responses:
        "200":
          description: Child removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "403":
          description: Only the owner can delete
  /children/{childId}/alerts:
    parameters:
      - $ref: "#/components/parameters/ChildId"
    get:
      tags: [Children]
      summary: List alerts for a child
      operationId: listChildAlerts
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Paginated alerts
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Alert"
                      pagination:
                        $ref: "#/components/schemas/Pagination"
  /children/{childId}/alerts/ack:
    parameters:
      - $ref: "#/components/parameters/ChildId"
    post:
      tags: [Children]
      summary: Acknowledge alerts
      operationId: acknowledgeAlerts
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                alert_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        "200":
          description: Alerts acknowledged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
  /children/{childId}/guardians:
    parameters:
      - $ref: "#/components/parameters/ChildId"
    get:
      tags: [Guardians]
      summary: List guardians
      operationId: listGuardians
      responses:
        "200":
          description: List of guardians
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Guardian"
    post:
      tags: [Guardians]
      summary: Add a guardian
      operationId: addGuardian
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        "201":
          description: Guardian added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "403":
          description: Forbidden
        "404":
          description: User not found
  /children/{childId}/guardians/{guardianId}:
    parameters:
      - $ref: "#/components/parameters/ChildId"
      - $ref: "#/components/parameters/GuardianId"
    delete:
      tags: [Guardians]
      summary: Remove a guardian
      operationId: removeGuardian
      responses:
        "200":
          description: Guardian removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "403":
          description: Forbidden
        "409":
          description: Conflict (e.g. cannot remove owner)
...
  /children/{childId}/screen-time-limit:
    parameters:
      - $ref: "#/components/parameters/ChildId"
    put:
      tags: [Screen Time]
      summary: Set daily screen time limit (0-1440 minutes, null to remove)
      operationId: setScreenTimeLimit
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [limit_minutes]
              properties:
                limit_minutes:
                  type: integer
                  minimum: 0
                  maximum: 1440
                  nullable: true
      responses:
        "200":
          description: Limit updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
...
  /children/{childId}/geofences:
    parameters:
      - $ref: "#/components/parameters/ChildId"
    get:
      tags: [Geofencing]
      summary: List geofences for a child
      operationId: listGeofences
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: List of geofences
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Geofence"
                      pagination:
                        $ref: "#/components/schemas/Pagination"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
    post:
      tags: [Geofencing]
      summary: Create a geofence
      operationId: createGeofence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGeofenceInput"
      responses:
        "201":
          description: Geofence created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "422":
          description: Validation error
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /children/{childId}/geofences/{geofenceId}:
    parameters:
      - $ref: "#/components/parameters/ChildId"
      - name: geofenceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    put:
      tags: [Geofencing]
      summary: Update a geofence
      operationId: updateGeofence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGeofenceInput"
      responses:
        "200":
          description: Geofence updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "422":
          description: Validation error
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Geofence not found
    delete:
      tags: [Geofencing]
      summary: Delete a geofence
      operationId: deleteGeofence
      responses:
        "200":
          description: Geofence deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Geofence not found
  /devices/{deviceId}/geofences/check:
    parameters:
      - $ref: "#/components/parameters/DeviceId"
    post:
      tags: [Geofencing]
      summary: Check device location against all active geofences (device-side)
      operationId: checkGeofences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [latitude, longitude]
              properties:
                latitude:
                  type: number
                  format: double
                longitude:
                  type: number
                  format: double
      responses:
        "200":
          description: Geofence check results
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          violations:
                            type: array
                            items:
                              $ref: "#/components/schemas/GeofenceViolation"
        "401":
          description: Unauthorized
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    sessionCookie:
      type: apiKey
      in: cookie
      name: kavach_access
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    ChildId:
      name: childId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    DeviceId:
      name: deviceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    GuardianId:
      name: guardianId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    Envelope:
      type: object
      properties:
        success:
          type: boolean
        data:
          oneOf:
            - type: object
            - type: array
            - type: string
            - type: number
            - type: boolean
            - type: "null"
          nullable: true
        error:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          format: uuid
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer
...
    Child:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        birth_date:
          type: string
          format: date
          nullable: true
    Alert:
      type: object
      properties:
        id:
          type: string
          format: uuid
        alert_type:
          type: string
        details:
          type: string
        acknowledged_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    Guardian:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        is_owner:
          type: boolean
        joined_at:
          type: string
          format: date-time
...
    Geofence:
      type: object
      properties:
        id:
          type: string
          format: uuid
        child_id:
          type: string
          format: uuid
        name:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        radius_meters:
          type: number
        zone_type:
          type: string
          enum: [HOME, SCHOOL, FRIEND, CUSTOM]
        alert_on_entry:
          type: boolean
        alert_on_exit:
          type: boolean
        device_id:
          type: string
          format: uuid
          nullable: true
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
    CreateGeofenceInput:
      type: object
      required: [name, latitude, longitude, radius_meters]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        latitude:
          type: number
          minimum: -90
          maximum: 90
        longitude:
          type: number
          minimum: -180
          maximum: 180
        radius_meters:
          type: number
          minimum: 50
          maximum: 50000
        zone_type:
          type: string
          enum: [HOME, SCHOOL, FRIEND, CUSTOM]
          default: CUSTOM
        alert_on_entry:
          type: boolean
          default: false
        alert_on_exit:
          type: boolean
          default: true
        device_id:
          type: string
          format: uuid
        is_active:
          type: boolean
          default: true
    UpdateGeofenceInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        latitude:
          type: number
          minimum: -90
          maximum: 90
        longitude:
          type: number
          minimum: -180
          maximum: 180
        radius_meters:
          type: number
          minimum: 50
          maximum: 50000
        zone_type:
          type: string
          enum: [HOME, SCHOOL, FRIEND, CUSTOM]
        alert_on_entry:
          type: boolean
        alert_on_exit:
          type: boolean
        is_active:
          type: boolean
    GeofenceViolation:
      type: object
      properties:
        geofence_id:
          type: string
          format: uuid
        name:
          type: string
        zone_type:
          type: string
        event_type:
          type: string
          enum: [ENTRY, EXIT]
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        recorded_at:
          type: string
          format: date-time
...